// functions/src/domain/match_reports/incident_validation_code.enum.ts
/** Reasons an incident fails validation. */
export enum IncidentValidationCode {
  INCIDENT_ID_REQUIRED = 'INCIDENT_ID_REQUIRED',
  IDEMPOTENCY_KEY_REQUIRED = 'IDEMPOTENCY_KEY_REQUIRED',
  IDEMPOTENCY_KEY_TOO_LONG = 'IDEMPOTENCY_KEY_TOO_LONG',
  INVALID_INCIDENT_TYPE = 'INVALID_INCIDENT_TYPE',
  INVALID_TEAM_SIDE = 'INVALID_TEAM_SIDE',
  TEAM_SIDE_REQUIRED = 'TEAM_SIDE_REQUIRED',
  INVALID_JERSEY_NUMBER = 'INVALID_JERSEY_NUMBER',
  INVALID_MINUTE = 'INVALID_MINUTE',
  NOTES_TOO_LONG = 'NOTES_TOO_LONG',
}